--- cg_lims/adapter/__init__.py
@@ -1,1 +1,0 @@
-from .plugin import CgLimsAdapter
