--- cg_lims/app/api/api_v1/endpoints/__init__.py
@@ -1,1 +1,0 @@
-
