--- cg_lims/app/api/deps.py
@@ -1,6 +1,0 @@
-from genologics.lims import Lims
-from genologics.config import BASEURI, USERNAME, PASSWORD
-
-
-def get_lims():
-    return Lims(BASEURI, USERNAME, PASSWORD)
