+++ cg_lims/commands/base.py
@@ -8,7 +8,6 @@ from cg_lims.status_db_api import StatusDBAPI
 
 # commands
 from cg_lims.EPPs import epps
-from cg_lims.commands.serve import serve_command
 
 
 @click.group(invoke_without_command=True)
@@ -24,11 +23,8 @@ def cli(ctx, config):
     ctx.obj["lims"] = lims
     ctx.obj["status_db"] = status_db
     ctx.obj["arnold_host"] = config_data.get("ARNOLD_HOST")
-    ctx.obj["host"] = config_data.get("CG_LIMS_HOST")
-    ctx.obj["port"] = config_data.get("CG_LIMS_PORT")
     ctx.obj["db_uri"] = config_data.get("DB_URI")
     ctx.obj["db_name"] = config_data.get("DB_NAME")
 
 
 cli.add_command(epps)
-cli.add_command(serve_command)
