+++ cg_lims/exceptions.py
@@ -1,3 +1,8 @@
+from typing import Optional
+
+from starlette import status
+
+
 class LimsError(Exception):
     """Base Error"""
 
@@ -82,3 +87,10 @@ class CSVColumnError(LimsError):
     """Raise when handling errors with csv columns"""
 
     pass
+
+
+class InsertError(LimsError):
+    def __init__(self, message: str, code: Optional[int] = status.HTTP_405_METHOD_NOT_ALLOWED):
+        self.message = message
+        self.code = code
+        super().__init__(message)
