+++ cg_lims/models/mongo/prep/wgs_prep.py
@@ -2,7 +2,7 @@ from typing import Optional
 from pydantic.main import BaseModel
 from pydantic import Field
 
-from cg_lims.models.database.prep import Prep
+from cg_lims.models.mongo.prep import Prep
 
 
 class BufferExchangeArtifactUDF(BaseModel):
@@ -26,7 +26,7 @@ class NormalizationOfMicrobialSamplesProcessUDFS(BaseModel):
     )
 
 
-class MicrobialLibraryPrepNexteraProcessUDFS(BaseModel):
+class MicrobialLibraryPrepNexteraProcessUDFS:
     lot_nr_tagmentation_buffer: str = Field(..., alias="Lot nr: Tagmentation buffer (TD-buffer)")
     lot_nr_tagmentation_enzyme: str = Field(..., alias="Lot nr: Tagmentation enzyme (TDE1)")
     lot_nr_index: str = Field(..., alias="Lot nr: Index")
@@ -69,7 +69,5 @@ class MicrobialPrep(
     BufferExchangeProcessUDFS,
     BufferExchangeArtifactUDF,
 ):
-    workflow = "Microbial"
-
     class Config:
         allow_population_by_field_name = True
